+++ src/main/java/wad/seoul_nolgoat/auth/oauth2/security/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.auth.oauth2;
+package wad.seoul_nolgoat.auth.oauth2.security;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
